% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_excel_versions.R
\name{compare_excel_files}
\alias{compare_excel_files}
\title{Title Compare tro excel files and gives the differences in Column names and rows}
\usage{
compare_excel_files(
  old_version_path,
  new_version_path,
  output_file,
  OVER_WRITE = TRUE
)
}
\arguments{
\item{old_version_path}{File path to old version}

\item{new_version_path}{File path to new version}

\item{output_file}{Path to wite the output file with xlsx extension}

\item{OVER_WRITE}{TRUE if need to overwrite the old version of output file}
}
\value{

}
\description{
Title Compare tro excel files and gives the differences in Column names and rows
}
\examples{
\dontrun old_version_path <- "K:/QILT/GOS/2023/Overall/10. Outputs/Data files/Specs/archive/GOS 2022 Master Output Spec.xlsx"
}
