% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_files.R
\name{find_files}
\alias{find_files}
\title{Title Find a file in a folder}
\usage{
find_files(
  folder_path,
  subfolder_path,
  file_name,
  file_extension,
  match_all = TRUE,
  strict = FALSE
)
}
\arguments{
\item{folder_path}{Path to the folder in which the file exists}

\item{subfolder_path}{Sub-folder name in any order}

\item{file_name}{A string from the name of the file}

\item{file_extension}{File extension for example xlsx or txt}

\item{match_all}{TRUE if all the values in file_name must be there in the output file.}

\item{strict}{If TRUE only gives results when all values in subfolder_path match}
}
\value{

}
\description{
Title Find a file in a folder
}
\examples{
find_files(folder_path = "K:/QILT/GOS/2023", subfolder_path = c("sample", "feb"), file_extension = "xlsx", file_name = c("Spec", "custom"), match_all = TRUE, strict = TRUE)
}
