% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lkp_vars.R
\name{lkp_var}
\alias{lkp_var}
\title{Sample varaibles, its definition and values}
\usage{
lkp_var(var)
}
\arguments{
\item{var}{A GOS variable name}
}
\value{
The definition of a variable, values and labels
}
\description{
Sample varaibles, its definition and values
}
\examples{
lkp_var("DOMINT")
}
