% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{sample_emails}
\alias{sample_emails}
\title{Helper to check email mis-sort in sample processing}
\usage{
sample_emails(df, ..., n = 10)
}
\arguments{
\item{df}{A dataframe to extract variables from (Sample file)}

\item{...}{Any other variable you want to display}

\item{n}{Number of samples}

\item{E402}{Last name}

\item{E403}{First name}

\item{EMAIL1}{Email1}

\item{EMAIL2}{Email2}

\item{EMAIL3}{Email3}
}
\value{

}
\description{
Helper to check email mis-sort in sample processing
}
