% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcsi_funs.R
\name{tcsi_var}
\alias{tcsi_var}
\title{QILT TCSI functions
These functions provide format checks for TCSI variables}
\usage{
tcsi_var(elm)
}
\value{

}
\description{
QILT TCSI functions
These functions provide format checks for TCSI variables
}
