% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lkp_vars.R
\name{tcsi_web}
\alias{tcsi_web}
\title{QILT TCSI Website Functions}
\usage{
tcsi_web(elm)
}
\arguments{
\item{elm}{A TCSI element}
}
\value{
OPen TCSI website for the element
}
\description{
QILT TCSI Website Functions
}
\examples{
tcsi_web("470")
}
