% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_files.R
\name{find_files}
\alias{find_files}
\title{To find a file in QILT project folders}
\usage{
find_files(
  project,
  year,
  drive = "",
  subfolder_path = "",
  file_name = "",
  file_extension = "",
  match_all_name = TRUE,
  match_all_path = FALSE
)
}
\arguments{
\item{project}{Specify the QILT project (GOS, SES, ESS, GOS-L)}

\item{year}{Specify the collection year}

\item{subfolder_path}{Sub-folder name of the file path in any order}

\item{file_name}{One or more keywords from the name of the file}

\item{file_extension}{File extension for example xlsx or txt}

\item{match_all_name}{TRUE if all the values in file_name must be there in the output file.}

\item{match_all_path}{If TRUE only gives results when all values in subfolder_path match}
}
\value{

}
\description{
To find a file in QILT project folders
}
\examples{
eval=FALSE
\dontrun{find_files(project = "GOS", year = "2023", subfolder_path = c("sample", "feb"), file_name = c("Spec", "sample"), file_extension = "xlsx", match_all_name = TRUE, match_all_path = TRUE)}
}
