% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lkp_vars.R
\name{lkp_var}
\alias{lkp_var}
\title{Sample variables, its definition and values}
\usage{
lkp_var(var)
}
\arguments{
\item{var}{A QILT variable name}
}
\value{
The definition of a variable, its values and labels
}
\description{
Sample variables, its definition and values
}
\examples{
eval = FALSE
\dontrun{lkp_var("e329")}
}
