% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcsi_funs.R
\name{tcsi_var}
\alias{tcsi_var}
\title{QILT TCSI functions}
\usage{
tcsi_var(elm)
}
\arguments{
\item{elm}{\itemize{
\item A TCSI element
}}
}
\value{

}
\description{
These functions provide format checks for TCSI variables
}
